fixed a portion of the problems
i fixed some problems. not all the problems, because some of them were too hard to fix, but i got most of them. to make myself feel better about the problems i couldn't fix, i added cool css to the boxes, so now it's harder to read the text but they fade from one color to another. also i added a title. i'm gonna stop working on this for now.